LPS-160218 Reorganise panel state utilities
