+++ modules/apps/content-dashboard/content-dashboard-web/types/src/main/resources/META-INF/resources/js/transformers/panelStateHandler.d.ts
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+declare const handlePanelStateFromSession: ({
+	currentRowId,
+	namespace,
+	panelState,
+	selectedItemFetchURL,
+	selectedItemRowId,
+}: {
+	currentRowId: string;
+	namespace: string;
+	panelState: string;
+	selectedItemFetchURL: string;
+	selectedItemRowId: string;
+}) => void;
+declare const handleSessionOnSidebarOpen: ({
+	panelState,
+	rowId,
+	selectedItemRowId,
+}: {
+	panelState: string;
+	rowId: string;
+	selectedItemRowId: string;
+}) => void;
+export {handlePanelStateFromSession, handleSessionOnSidebarOpen};
